% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oddsratio.R
\name{Oddsratio}
\alias{Oddsratio}
\title{Odds Ratio}
\usage{
Oddsratio(X, Y = NULL)
}
\arguments{
\item{X}{a n x 1 numeric vector, matrix or data frame. If a contingency table is supplied, the upper left corner shall contain the joint success probability (or frequency). If a 3-dimensional vector of probabilities is supplied, the order (p, q, r) shall be respected.}

\item{Y}{NULL (default) or a n x 1 numeric vector, matrix or data frame and compatible dimensions to X.}
}
\value{
The value of the oddsratio
}
\description{
\code{Oddsratio()} computes the oddsratio
}
\examples{
X <- matrix(c(10, 20, 30, 5), ncol = 2)
Oddsratio(X)
}
