% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2x2_inverse.R
\name{YuleQ.inv}
\alias{YuleQ.inv}
\title{Inverse Yule's Q for 2 x 2 contingency tables}
\usage{
YuleQ.inv(Q, m)
}
\arguments{
\item{Q}{Yule's Q}

\item{m}{a 4 x 1 vector of marginals c(R1, R2, C1, C2)}
}
\value{
a matrix corresponding to the inputs of a 2x2 contingency table
}
\description{
\code{YuleQ.inv()} computes a 2 x 2 contingency table based on the marginal frequencies or probabilities and Yule's Q.
}
\examples{
Q <- 0.5
m <- c(10, 40, 20, 30)
YuleQ.inv(Q, m)
}
\references{
\itemize{
\item \insertRef{pohle2024measuringdependenceevents}{BCor}
\item \insertRef{yule1900}{BCor}
}
}
