% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YuleQ.R
\name{YuleQ}
\alias{YuleQ}
\title{Yule's Q}
\usage{
YuleQ(X, Y = NULL, g = 1, alpha = 0.95, Fisher = TRUE, covar = "iid", n)
}
\arguments{
\item{X}{a n x 1 numeric vector, matrix or data frame with values of 0 or 1. If a contingency table is supplied, the upper left corner shall contain the joint success probability (or frequency). If a 3-dimensional vector of probabilities is supplied, the order (p, q, r) shall be respected.}

\item{Y}{NULL (default) or a n x 1 numeric vector, matrix or data frame with values of 0 or 1 and compatible dimensions to X.}

\item{g}{a scalar between 0 and 1.}

\item{alpha}{confidence level for the returned confidence interval. FALSE yields Yule's Q without confidence intervals.}

\item{Fisher}{Indicator whether confidence intervals should be computed by using the Fisher Transformation. Default is TRUE.}

\item{covar}{data assumptions: iid ("iid"), heteroskedasticity ("HC") or heteroskedasticity and autocorrelation ("HAC").}

\item{n}{sample size. Only necessary if probabilities are provided and confidence intervals are desired.}
}
\value{
The value of Yule's Q or any of its related measures. g = 0.5 yields Yule's Y and g = 0.75 yields Digby's H.
}
\description{
\code{YuleQ()} computes Yule's Q and corresponding confidence intervals.
}
\examples{
x <- matrix(c(10, 20, 30, 5), ncol = 2)
YuleQ(x)
}
