% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2x2_inverse.R
\name{Cole.inv}
\alias{Cole.inv}
\title{Inverse Cole's C for 2 x 2 contingency tables}
\usage{
Cole.inv(C, m)
}
\arguments{
\item{C}{Cole's C}

\item{m}{a 4 x 1 vector of marginals c(R1, R2, C1, C2)}
}
\value{
a matrix corresponding to the inputs of a 2x2 contingency table
}
\description{
\code{Cole.inv()} computes a 2 x 2 contingency table based on the marginal frequencies or probabilities and Cole's C.
}
\examples{
C <- 0.5
m <- c(10, 40, 20, 30)
Cole.inv(C, m)
}
\references{
\itemize{
\item \insertRef{pohle2024measuringdependenceevents}{BCor}
\item \insertRef{Cole1949}{BCor}
}
}
