% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cole.R
\name{Cole}
\alias{Cole}
\title{Cole's Correlation}
\usage{
Cole(X, Y = NULL, alpha = 0.95, covar = "iid", m_rep = 10000, c_seq = NA)
}
\arguments{
\item{X}{a n x 1 numeric vector, matrix or data frame.}

\item{Y}{a n x 1 numeric vector, matrix or data frame.}

\item{alpha}{confidence level for the returned confidence interval.}

\item{covar}{data assumptions: iid ("iid"), heteroskedasticity ("HC") or heteroskedasticity and autocorrelation ("HAC").}

\item{m_rep}{number of Monte Carlo replications used for approximating the limiting distribution of Cole's C.}

\item{c_seq}{sequence of C_0 to be tested for computing the confidence interval. Default is the sequence from -0.999 to 0.999 with distance 0.001.}
}
\value{
The value of Cole's correlation.
}
\description{
\code{Cole()} computes Cole's correlation and corresponding confidence intervals.
}
\examples{
x <- matrix(c(10, 20, 30, 5), ncol = 2)
Cole(x)
}
