% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tetrachoric.R
\name{Tetrachoric}
\alias{Tetrachoric}
\title{Tetrachoric Correlation}
\usage{
Tetrachoric(X, Y = NULL)
}
\arguments{
\item{X}{a n x 1 numeric vector, matrix or data frame.}

\item{Y}{a n x 1 numeric vector, matrix or data frame.}
}
\value{
The value of the tetrachoric correlation.
}
\description{
\code{Tetrachoric()} computes the tetrachoric correlation.
}
\examples{
X <- matrix(c(10, 20, 30, 5), ncol = 2)
Tetrachoric(X)
}
