% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2x2_inverse.R
\name{YuleQ.inv}
\alias{YuleQ.inv}
\title{YuleQ.inv}
\usage{
YuleQ.inv(Q, m)
}
\arguments{
\item{Q}{Yule's Q}

\item{m}{a 4 x 1 vector of marginals c(R1, R2, C1, C2)}
}
\value{
a matrix corresponding to the inputs of a 2x2 contingency table
}
\description{
YuleQ.inv
}
\examples{
Q <- 0.5
m <- c(10, 40, 20, 30)
YuleQ.inv(Q, m)
}
