% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YuleQ.R
\name{YuleQ}
\alias{YuleQ}
\title{Yule's Q}
\usage{
YuleQ(
  X,
  Y = NULL,
  g = 1,
  alpha = 0.95,
  Fisher = TRUE,
  covar = "iid",
  n = 10000
)
}
\arguments{
\item{X}{a n x 1 numeric vector, matrix or data frame with values of 0 or 1. If a contingency table is supplied, the upper left corner shall contain the joint success probability (or frequency). If a 3-dimensional vector of probabilities is supplied, the order (p, q, r) shall be respected.}

\item{Y}{NULL (default) or a n x 1 numeric vector, matrix or data frame with values of 0 or 1 and compatible dimensions to X.}

\item{g}{a scalar between 0 and 1.}

\item{alpha}{confidence level for the returned confidence interval. FALSE yields Yule's Q without confidence intervals.}

\item{Fisher}{Indicator whether confidence intervals should be computed by using the Fisher Transformation. Default is TRUE.}

\item{covar}{method that is used to compute the covariance matrix, denoted Omega in \insertCite{pohle2024measuringdependenceevents}{BCor}, see in particular Appendix C. Options are eiher  iid ("iid"), heteroskedasticity ("HC") or heteroskedasticity and autocorrelation ("HAC“), where the latter two follow the method of \insertCite{NeweyWest1987}{BCor}.}

\item{n}{sample size. Only necessary if probabilities are provided and confidence intervals are desired.}
}
\value{
The value of Yule's Q or any of its related measures. g = 0.5 yields Yule's Y and g = 0.75 yields Digby's H. Confidence intervals are only implemented for Yule's Q (g = 1).
}
\description{
\code{YuleQ()} computes (generalized) Yule's Q and corresponding confidence intervals.
}
\examples{
# Insert a contingency table with frequencies in form of a matrix.
x <- matrix(c(10, 20, 30, 5), ncol = 2)
YuleQ(x)

# Insert a contingency table with relative frequencies in form of a matrix.
# In that case, disable confidence intervals via alpha = FALSE or supply a sample size n.
x <- matrix(c(0.2, 0.1, 0.4, 0.3), ncol = 2)
YuleQ(x, alpha = FALSE)

# Insert two vectors of observations.
x <- c(0,1,1,1,1,0,1,0,0,1)
y <- c(1,0,1,1,0,0,0,0,1,0)
YuleQ(x, y)

# Insert two marginal success probabilities (p for the row variable and q for the column variable)
# and a joint success probability r. In that case, disable confidence intervals via alpha = FALSE
# or supply a sample size n.
p <- 0.6
q <- 0.3
r <- 0.2
YuleQ(c(p, q, r), alpha = FALSE)
}
\references{
\itemize{
\item \insertRef{pohle2024measuringdependenceevents}{BCor}
\item \insertRef{yule1900}{BCor}
\item \insertRef{yule1912}{BCor}
\item \insertRef{NeweyWest1987}{BCor}
}
}
