% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YuleQ.R
\name{YuleQ}
\alias{YuleQ}
\title{Yule's Q}
\usage{
YuleQ(X, Y = NULL, g = 1, alpha = 0.95, covar = "iid")
}
\arguments{
\item{X}{a n x 1 numeric vector, matrix or data frame.}

\item{Y}{a n x 1 numeric vector, matrix or data frame.}

\item{g}{a scalar between 0 and 1.}

\item{alpha}{confidence level for the returned confidence interval.}

\item{covar}{data assumptions: iid ("iid"), heteroskedasticity ("HC") or heteroskedasticity and autocorrelation ("HAC").}
}
\value{
The value of Yule's Q or any of its related measures. g = 0.5 yields Yule's Y and g = 0.75 yields Digby's H.
}
\description{
\code{YuleQ()} computes Yule's Q and corresponding confidence intervals.
}
\examples{
x <- matrix(c(10, 20, 30, 5), ncol = 2)
YuleQ(x)
}
